requirements.txt
mss>=9.0.1
# pynput must be installed without evdev on Linux (evdev needs python3-devel to compile):
#   pip install python-xlib && pip install pynput --no-deps
# On Windows/macOS: pip install pynput
python-xlib>=0.33
pynput>=1.7.7
Pillow>=10.0.0
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
rich>=13.0.0
